requirements.txt
numpy
torch
torchvision
ruamel.yaml
matplotlib